refactor: Move infrastructure file into right directory
Otherwise it detracts from the test coverage.
